+++ app/src/main/java/com/github/amusinamaria/repository/CastCard.kt
@@ -1,4 +1,4 @@
-package com.example.androidacademymsk.repository
+package com.github.amusinamaria.repository
 
 data class CastCard(
     val name: String,
