+++ app/src/main/java/com/github/amusinamaria/repository/MockDataSource.kt
@@ -1,4 +1,4 @@
-package com.example.androidacademymsk.repository
+package com.github.amusinamaria.repository
 
 class MockDataSource {
     fun getMovieCards(): List<MovieCard> {
