+++ app/src/main/java/com/github/amusinamaria/ui/details/CastAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.androidacademymsk.ui.details
+package com.github.amusinamaria.ui.details
 
 import android.view.LayoutInflater
 import android.view.View
@@ -11,8 +11,8 @@ import com.bumptech.glide.load.MultiTransformation
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.bumptech.glide.request.RequestOptions
-import com.example.androidacademymsk.R
-import com.example.androidacademymsk.repository.CastCard
+import com.github.amusinamaria.R
+import com.github.amusinamaria.repository.CastCard
 import jp.wasabeef.glide.transformations.RoundedCornersTransformation
 
 class CastAdapter : RecyclerView.Adapter<CastAdapter.CastViewHolder>() {
