+++ app/src/test/java/com/github/amusinamaria/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.example.androidacademymsk
+package com.github.amusinamaria
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
